fastapi/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
langchain-community
langgraph
pydantic
pydantic-settings
python-dotenv
requests
httpx
websockets
langsmith
tavily-python
ortools
googlemaps
